initialize the Docker client in a prerun
